kajiya: Ensure all files can be deleted during sandbox deletion
Actions or side-effects of the sandboxing mechanism can cause
directories inside the sandbox tree to not have write access, which
causes os.RemoveAll to fail. By explicitly fixing the permissions of all
directories in the sandbox with a call to os.Chmod, we can ensure that
we will always be able to successfully delete the sandbox.

Change-Id: I793c2b07f7ae8268169a12e4571ea70fd213cb3e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5117087
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61956}