[tricium] Fetch shellcheck to pin from go/subtool/shellcheck instead of SDK.
http://go/subtool/shellcheck is a CIPD package created using the SDK as a toolchain, but distinct from the core SDK builder. It updates
automatically when the shellcheck ebuild is revbumped.

This will allow the shellcheck dependency to be dropped from
virtual/target-chromium-os-sdk-nobdeps (it remains in
virtual/target-sdk-subtools).

Migrate the scripts, and update paths, to fetch and use shellcheck
from the subtool CIPD package rather than the prebuilt SDK. Note the
`shellcheck_tricium` CIPD package is separate, and repackages the
subtool. Changes to this package described in `cipd.yaml` are tested,
locally but not uploaded.

The version used is pinned in an ensure file. This CL updates the pin in the old scripts from `0.8.0-r55` to
`ebuild_source:dev-util/shellcheck-0.8.0-r65` which is the first
available in the new prefix. Precisely,
http://go/cipd/p/chromiumos/infra/tools/shellcheck/+/egOJJytEgvaYBWkwzgea8ERPILo-3qRH64DB6Fa72doC

Remaining steps (per README.md):

  $ cipd create -pkg-def cipd.yaml
  # outputs <VERSION>
  $ cipd set-ref infra/tricium/function/shellcheck -ref live -version <VERSION>

Bug: b/277662650
Test: make testrun
Test: cipd pkg-build -pkg-def cipd.yaml -out pkg.zip && unzip...
Change-Id: Ic386244a9ec5cf8462cf059b2d77fa5bbbb1b58b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5124787
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/main@{#61984}