[codeql] Add the CodeQL CLI to infra/3pp.
We're creating a buidler + recipe to create a CodeQL database of the
Chromium codebase on a weekly basis.

In order for that recipe to work, the script at
//src/tools/codeql/index_target.py expects to have access to the CodeQL
command line tool.

Adding this dependency to infra/3pp will allow the script access to this
tool on the builders.

Bug: 1506812
Change-Id: Ie2b4f2bc19bb214e96f9a1057dfa0058998d46cc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5123886
Reviewed-by: Brian Ryner <bryner@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#62001}