Remove xinyuoffline@google.com from go/src/infra/isolation/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/xinyuoffline@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Icd8d5a87d4108cd161db686c22a46141f065cd3a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5150921
Reviewed-by: Suma Kasa <sumakasa@google.com>
Commit-Queue: Suma Kasa <sumakasa@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#62129}