Remove sfetisov@chromium.org from go/src/infra/cmd/drone-agent/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.
If no no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/sfetisov@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I0361395760957c64d6ac9c952773c9dd650217f6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5175398
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Xixuan Wu <xixuan@google.com>
Reviewed-by: Xixuan Wu <xixuan@google.com>
Cr-Commit-Position: refs/heads/main@{#62169}