pkg_index: Invert package-support checks
Based on https://crrev.com/c/5074781, it seems like package_sleuth is
trying to skip unsupported packages. But it seems like we're actually
doing the opposite.

IsPackageSupported() returns a PackageSupport IntEnum, which defines
PackageSupport.SUPPORTED = 0. Thus, if you're checking the truthiness of
IsPackageSupported() like we do in package_sleuth, you'll get False iff
the package is supported. This is backward.

To prevent the likelihood of future bugs like this, I've added two new
methods to PackageSupport: is_supported() and is_unsupported().
Hopefully future devs will check those functions rather than implicitly
checking the truthiness of a PackageSupport instance.

I've also renamed IsPackageSupported() to GetPackageSupport(), which
should clarify that it's returning a PackageSupport, not a bool.

Finally, I've replaced all of our "is this package supported?" checks
with calls to is_supported() and is_unsupported().

BUG=b:319258767
TEST=call script, passing in a supported package; make sure it runs.
Note that this actually outputs a different set of "packages going
forward" than the stdout we see in a real build, since we're now
crawling the depgraph based on what's supported, instead of what's
unsupported.
TEST=call script with an unsupported package; make sure nothing happens.

Change-Id: I3665b151dce0c32f4f9aba4c05291cacb6bde58e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5182575
Commit-Queue: George Engelbrecht <engeg@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>
Cr-Commit-Position: refs/heads/main@{#62223}