satlab: Switch MaxTimeout->TimeoutMins
Our builders (temporarily) have different timeouts, and additionally
these do not line up with the main lab timeouts. Therefore a global
`max-timeout` is not applicable, and may not be in the future.

For users of the UI, the experience will remain the same (timeout gets
set to 2370 mins, arbitrarily high), while for the CLI a user will need
to set `-timeout-mins`. If a user is interfacing with the CLI, it is
assumed they will be sophisticated enough to set reasonable values.

Change-Id: If840f93c9eee78b4b1e38ec982e00a6299f0ef71
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5184980
Auto-Submit: Eli Trexler <elijahtrexler@google.com>
Commit-Queue: Varun Thammineni <varunth@google.com>
Reviewed-by: Varun Thammineni <varunth@google.com>
Cr-Commit-Position: refs/heads/main@{#62265}