[Dev] Add a dynamic bucket in Dev to test Taskbackendlite service
Add a dynamic bucket in Dev to test Taskbackendlite service. This bucket
config is similar to the WebRTC one. It doesn't specify any recipe to
run. The expected behavior is that after it's manually triggered in
luci-scheduler-dev UI, we could see the pubsub msg in the corresponding
taskbackendlite topic. After `heartbeat_timeout`, this build should be
marked a failure.

Bug: 1502021
Change-Id: I32b9550a9621b95df6e00304a9c976767d1cd5ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5182258
Commit-Queue: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Randy Maldonado <randymaldonado@google.com>
Cr-Commit-Position: refs/heads/main@{#62277}