result_adapter: add non-verbose mode for go tests
This change adds a flag to the "go" subcommand of result_adapter which
copies test output to stdout in non-verbose mode. Non-verbose mode is
intended to replicate the output of "go test" without -v. For projects
with a lot of tests, this dramatically reduces the size of the output by
hiding all the test logs for passing tests, and summing up passing
packages with a single line.

While we're here, since we need to accumulate output for each test and
package in non-verbose mode anyway, make the standard verbose mode's
output cleaner. Currently, if packages are tested in parallel, the
output will have confusingly interleaved output. We can use the same
logic to emit the verbose test output grouped by package and test,
making it more understandable.

Credit to austin@google.com for writing most of the translation of JSON
back to non-verbose mode.

Change-Id: I6b3a4ee73eebaa9a8f282d66dc0173f4e34aeddb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5193848
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62354}