[golangbuild] run "result_adapter go" in non-verbose mode
This change changes the invocation of "result_adapter go" to include the
"-v=false" flag, which sets the output of the command to more closely
match "go test" in non-verbose mode.

This is useful both for human readability (the default verbose output
is often way too long to skim) and for watchflakes log-level match
patterns, which rely on the non-verbose output.

Change-Id: I16474980343beae038539c2e7200780c1d557f6b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5201456
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62404}