[golangbuild] refactor getGo to not depend on buildSpec
This change refactors getGo and its callees to not depend on having a
complete buildSpec. The reason we want this is that a new builder mode
for performance testing, which will be added in the very near future,
will need to check out more than one Go toolchain. It'll need to do
basically exactly the same thing getGo does, but with a different
destination and source. So, this CL rewrites getGo to not depend on
buildSpec, but offers a convenience function, getGoFromSpec, that does
exactly what getGo used to do.

This sounds simple on the surface, but the main issue here is that the
dependency on buildSpec has trickled down into getGo's callees. The most
difficult to resolve parts of this involve the fact that the CAS
instance and tools root are coupled to the buildSpec. To work around
this, the CL stores the CAS instance and the tools root in the context
instead. This makes some sense too since it's more a property of the
environment on the machine than a property of the build per se, so it
seems OK to do things this way for now.

Change-Id: Iddd00d187f9b033e653901a37931c9b16711df55
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5191128
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62440}