[UFS] CreateMachineLSE fail on request to create empty lse
This patch changes the behaviour of the CreateMachineLSE API to fail
creating the MachineLSE record if the LSE is not one of the recognized
types. This is required to ensure the integrity of the database as a lot
of the other API assume that the record is of some specific type and
don't take into account a situation where it's of no type (nil).

Bug: b/295073214, b/320670551

Change-Id: Ia32af378fdd5c87890163694688f5deb1e9e8bc5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5202486
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62441}