[golangbuild] add support for perf comparisons with parent commits
This change adds support for the "parent" string to the baseline
property for MODE_PERF builds. What this enables is checking the
performance delta of a single commit, without having to know what the
commit's parent is.

This is going to be very useful for checking the performance impact of
CLs in presubmit, because this is a comparison that is made often.
(We'll also have a builder in presubmit for checking against the tip of
the current branch, so stacks of CLs will work too.)

Change-Id: I0038729b20e7eae9911df55b5b7e6a16f703a825
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5189153
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Cr-Commit-Position: refs/heads/main@{#62446}