[golangbuild] summarize benchmark results in perf mode
This change adds a benchmark results summary log to the top-level logs
of the build, produced by benchstat. This CL modifies golangbuild to
fetch benchstat from CIPD for perf mode builds, and then runs benchstat
against the output from cmd/bench. It then places this in a top-level
log (build.State.Log) and adds tracking for these top-level logs in
a logger value in the current context.Context, so that they can be
rendered in the top-level summary markdown.

All of this is going to come in handy for not only inspecting these
builds by hand, but also for allowing the Go team to run performance
tests against their in-progress CLs and see useful results.

Change-Id: I2594186fc71abf128afc4bc5398cf60aab6d4c98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5191130
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62481}