[test-health] Include target_platform for correction join
There are duplicate rows in the raw table. This looks to be caused by
the merge with the correction data when the target_platform changes.
Also group by component to avoid a change in components in a day having
unpredictable consequences

Bug: 1435578
Change-Id: I219dd62e78b96abebef9c6a3586177b27651c361
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5225564
Reviewed-by: Gary Tong <gatong@chromium.org>
Commit-Queue: Gary Tong <gatong@chromium.org>
Auto-Submit: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#62544}