sync_generated_files: Change opts.ignore type
Previously, I had the arg set to default=set() with action='append'. But
sets don't have an append method, so that is causing errors today.

Now the arg gets saved as a list, and then we cast it to a set before
doing actual work with it.

BUG=b:321040083
TEST=./recipes.py test train
TEST=run unit tests
TEST=run script with --ignore

Change-Id: I020df52b663aa622c37d3afc2eec497f6f913eeb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5240234
Commit-Queue: Josip Sokcevic <sokcevic@chromium.org>
Reviewed-by: Josip Sokcevic <sokcevic@chromium.org>
Auto-Submit: Greg Edelston <gredelston@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#62591}