upload_debug_symbols: Don't deduplicate native symbols
The symbol uploader API doesn't support checking for previously uploaded
symbols by symbol upload type.  The only symbols it considers are the
breakpad symbols so if we've already uploaded the breakpad symbols we
won't know that the native symbols are missing. Rejigger the logic in
the splitdebug step so that the breakpad symbol file is placed next to
the native symbols in the same debug Id named directory. Then while
filtering the breakpad symbols, only add tasks to upload the native
symbols when the breakpad symbol is missing.

This allows us to use the uploader API to figure out if native symbols
have been uploaded already by indirectly querying the status of the
breakpad symbols. If the breakpad symbols are missing, then the native
symbols are also missing. If the breakpad symbols are there, the native
symbols might or might not be there, because we haven't been uploading
native symbols in the past on ChromeOS. Eventually this will upload all
the native symbols though, once the breakpad symbols all change debug
Id.

BUG=b:171334299
TEST=go test
TEST=./upload_debug_symbols upload -dry-run=true -gs-path=gs://chromeos-image-archive/brya-release/R117-15544.0.0 -data-type=splitdebug

Change-Id: Ibfb2d57fb003976bd12e54c77c40e250e4edbf94
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5202114
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62599}