[golangbuild] add property to override C compiler
This change adds the ability to override the C compiler used by the
build by adding a property to set the CC environment variable to some C
compiler in the tools installation.

The main target of this is to support switching out the C compiler in
our builds easily and specifying as much of that as possible in the
builder configuration itself. In particular, this is necessary to
support clang-based builders. We may want to consider moving the Windows
C compiler logic into this as well.

Change-Id: Ic01612691d551850812d0099a79f2fd735f62d57
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5245175
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#62634}