chromium_status: ensure key is always an integer
When the Status was being created from PRPC rather than data store, key
was empty, which was serialized as "".  This broke when LUCI Notify
tried to parse it as an int64.

Detect this condition in serialization and ensure a valid integer is
written out instead of a string.

Change-Id: I41f646754de26a80e861599e43501059aa0ab3fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5247727
Auto-Submit: Matthew Warton <mwarton@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#62650}