Add a GitUniversalSource wheel type and use it for expect_tests.
This requires a small change to the semantics of
concurrency.PROCESS_SPAWN_LOCK. Previously, this lock was held for the
entire duration that any file is open for write. However, to run 'git'
during the wheel source download, this needs to change. Since files
are marked as non-inheritable by default in Python 3, we only need to
deal with the short race conditions during file open where the file may
not yet have been marked non-inheritable. So, hold the lock across
opening the wheel source archive for write, but release it once the
open() call has returned.

expect_tests will be used to run infra tests with vpython3.

Bug: 1413668
Change-Id: I94d258e30ddac644067960493f473d297fb0413d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5247728
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#62664}