Remove seanmccullough@chromium.org from go/src/infra/appengine/tree-status/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/seanmccullough@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ie18d2ac338020c025c3c7affd404b8f98627de87
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5249532
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62672}