Remove anhdle@chromium.org from go/src/infra/cros/satlab/satlab/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/anhdle@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I2112f47edc66a26188d9fed3f09d913455bd5337
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5250927
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#62673}