Remove anhdle@chromium.org from go/src/infra/cros/tlslib/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/anhdle@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I6afd3a97bd8746272f171175df749ad1d1d9d0c2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5250220
Commit-Queue: Congbin Guo <guocb@chromium.org>
Reviewed-by: Congbin Guo <guocb@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62674}