Remove keybo@google.com from go/src/infra/cr_builder_health/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/keybo@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Idd884b70572b55fe4fa3df0dd4787654b775d5d9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5249821
Reviewed-by: Wanda Mora <morawand@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Wanda Mora <morawand@google.com>
Cr-Commit-Position: refs/heads/main@{#62675}