Remove cjacomet@google.com from go/src/infra/appengine/rubber-stamper/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/cjacomet@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Idbaff091cd88e696abf6e6c13809bc203ecb3dc5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5249819
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Joey Scarr <jsca@google.com>
Commit-Queue: Joey Scarr <jsca@google.com>
Cr-Commit-Position: refs/heads/main@{#62679}