Remove seanmccullough@chromium.org from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/seanmccullough@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I433a463ff3d9e7d1cb87a793ba217ef42f8be807
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5249820
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62691}