Remove seanmccullough@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/seanmccullough@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I657a48c7db49d876b4ae6c7b3b9e86c0212e8636