Remove anhdle@google.com from go/src/infra/cmd/stable_version2/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/anhdle@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I5cf7ec6a022ea727df78e367bf7e88f3815ac9fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5253062
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#62708}