Remove elijahtrexler@google.com from go/src/infra/cros/satlab/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/elijahtrexler@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Id549a5e678c3bcae53279275830b493a1d1c36c3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5252437
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Prasad Vuppalapu <prasadv@chromium.org>
Commit-Queue: Prasad Vuppalapu <prasadv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62709}