[golangbuild] set up clang tools in PATH
This change sets up clang tools in PATH. Currently the clang builders
are failing because the ASAN output doesn't get symbolized. I suspect
this is due to the fact that there's no way to find the LLVM symbolizer
that ships with clang releases.

Even if that isn't the case, setting up clang in PATH is more robust
anyway, and the CC environment variable looks more like what someone
would set locally, which is a positive.

Change-Id: Iebb4fad9055cd1526bb0710cedf9e75fce1c2713
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5254569
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62754}