[golangbuild] enable misc-compile coverage for linux-arm port
The current version of the linux-arm builder will have fixed capacity,
which is enough for post-submit and explicit SlowBot requests, but not
enough for a pre-submit builder included by default. So stop skipping
it inside misc-compile coverage. (It's the only first class port with
this GOARCH value.)

For go.dev/issue/65241.

Change-Id: Ia22955e712d3924ff3d30d7611e929f56926aa7d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5254800
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62757}