[cloudbuildhelper] Add helper for tracking dependencies.
We are partially reimplementing "go mod vendor" in order to be
able to package GAE apps as self-contained Go modules with
minimal dependencies all vendored in. Google Cloud GAE deployment
machinery essentially just passes such bundles ~= unmodified to
"go build ..." (unlike when running in GOPATH mode, where it tries
to do local deps analysis in a buggy way that broke our deployments
numerous times already).

We can't use "go mod vendor" for this, since it will be bundling
dependencies of **all** GAE apps in the repository, not just the
app we are interested in. Considering that infra.git and
luci-go.git have many-many projects in them, this would result
in a high number of unnecessary redeployments of everything on
every commit.

In order for go to build stuff stored in such minimal bundles it
needs a go.mod file with version numbers of external modules
(versions are essentially just embedded into the resulting
binary, they aren't used for any registry lookups), as well
as vendor/modules.txt that describes the layout of "vendor/..."
directory with source code of bundled external packages.

vendor/modules.txt is just a list of vendored packages
along with their versioning info (it must match go.mod,
it is used by Go to know if the vendor/* directory is
up-to-date).

The CL adds a helper that can generate both go.mod and modules.txt
based on a list of packages the main package transitively depends
on.

Note that modules.txt is undocumented. But it is supposed to be
checked-in into the source code when vendoring dependencies, so its
format will likely be stable enough across Go releases.

R=bryner@google.com, fancl@google.com

Change-Id: I8a67b28180856ef367dba73c50d6f9c274c06107
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5254228
Reviewed-by: Chenlin Fan <fancl@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62765}