Remove seanmccullough@google.com from OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/seanmccullough@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I657a48c7db49d876b4ae6c7b3b9e86c0212e8636
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5253905
Commit-Queue: Erik Staab <estaab@chromium.org>
Reviewed-by: Erik Staab <estaab@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Daniel Ochoa <ochoadaniel@google.com>
Cr-Commit-Position: refs/heads/main@{#62773}