recovery: Set cellular labels if missing
This configures the modem and SIM label update actions to be run
during repair if they have not already been populated on celluar DUTs.
This information is relatively expensive to query and is static
throughout the lifetime of the DUT so we only need to query it if it's
missing.

BUG=b:322556662
TEST=local repair
TEST=go test

Change-Id: Ied183d1dec8bcf8a499b7a8e429824020a39ec8a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5251428
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Jason Stanko <jstanko@google.com>
Cr-Commit-Position: refs/heads/main@{#62783}