Remove seanmccullough@chromium.org from go/src/infra/appengine/sheriff-o-matic/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/seanmccullough@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I8cf5e397308f4b24e963e61ae44c79657492fb5e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5250831
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#62828}