[golangbuild] add new perf mode baseline "latest_go_release"
This change adds the new baseline option for perf mode
"latest_go_release" which looks at the source spec that triggered the
build (for the main Go repo) and finds the latest stable release for
that branch. For example, a change to tip will pick Go 1.22.1 if that's
the latest release overall. A change to release-branch.go1.21 will
pick Go 1.21.7, if that's the latest minor release of Go 1.21.

This is the port of the logic our old infrastructure uses to pick
a baseline.

Change-Id: Ib3cf662cc2f13f09608c6921e4b926f38b24b8d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5245176
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Michael Pratt <mpratt@google.com>
Cr-Commit-Position: refs/heads/main@{#62849}