[golangbuild] add support for subrepos in perf mode
This change adds support for subrepos in perf mode build, following with
the old infrastructure did (but with a little bit less redundant work).

Subrepos like x/tools will also want their baseline repo version to be a
tag, so add support for tags as a baseline here. Adding support for tags
is a little tricky, because our sourceSpec abstraction requires always
having a branch. This change modifies sourceSpec to allow support for
specifying tags without a branch. This actually simplifies some of the
code around creating source specs, but it does complicate the
abstraction. To help with that, this change also adds more validation to
fetchRepo.

Change-Id: I4ead21a0d550f3cdefc8fdccde6bf01d1d7bad6c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5251409
Reviewed-by: Michael Pratt <mpratt@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62850}