kajiya: Implement containerized execution and sandboxing with nsjail
Actions can request to be executed inside a container, whose image they
specify as a docker:// URL in a "container-image" property in the
platform properties.

Kajiya will download this image and extract it into a content-addressed
image repository by calling out to `docker create` and `docker export`.

Docker has a high per-action overhead for execution, so instead of
calling `docker run` to execute actions, we implement our own
containerized execution using nsjail. This achieves high security
and high performance with almost zero startup overhead for the action.

Kajiya now uses nsjail even for actions that do not specify a container
image. In this case, we mount the host's `/` as the chroot of the jail.

Change-Id: I01b33c626588b76c1a02fcde8cfb9f4d6112b62d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5117088
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Reviewed-by: Junji Watanabe <jwata@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62882}