[golangbuild] ensure only branch or tag are set in sourceForRef
CL 5251409 erroneously assumed strings.CutPrefix returned an empty
string if it returned false. This is not true: it returns the original
string. This causes many subrepo builds to fail.

This is a quick fix, but we really need unit tests. Either that, or the
sourceSpec invariants need to be simpler.

Change-Id: I6ded1726e6bb8eac6ca87b1506ae5200adf50288
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5273007
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#62886}