[cloudbuildhelper] Add a hack to workaround GAE weirdness.
When using go.mod GAE deployment mode, with app.yaml in a different
directory than go.mod (which is the case for absolutely all our GAE
apps), the resulting runtime image has static and resource files
in a wrong directory, breaking their usage. This is a regression
compared to GOPATH mode. See b/323980048 for details.

We'll just copy all potential resources to the root of the module.
If they collide with something existing already (e.g. existing Go
code package, or perhaps already copied files from another bundle),
then bail with a cryptic error message: such files will have to be
renamed or something. We'll see if it happens for real.

Note that we copy (instead of moving) in hope that the bug will be
fixed one day and deployments will start using "correct" files.

Note also that for frontend heavy apps like Milo this significantly
increases the final image size. And this is on top of unavoidable
~2x increase when using go.mod GAE deployment mode to begin with
(for some unfathomable reason in this mode GAE builder keeps all
original Go sources in the final image, in addition to the
compiled binary; and barely documented/discoverable, but technically
existing, GOOGLE_CLEAR_SOURCE [1] appears to have no effect on that
whatsoever).

Appengine is fun.

[1]: https://cloud.google.com/docs/buildpacks/go#compilation_flags

R=bryner@google.com, fancl@google.com
B=b/323980048

Change-Id: I7d8da4edaa655ebafe1f6978803c76fb32e8864c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5272926
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#62917}