upload_debug_symbols: Upload splitdebug files as ELF type
Change the upload type from DEBUG_ONLY to ELF so that crash considers
the debuginfo as a full binary for the purposes of debugging crashes.
The .debug (splitdebug) files are actually the full binary, not just the
debuginfo for a binary split from the binary itself. This is because of
CL:1865517 which doesn't do any objcopy splitting.

Even without that CL though, it seems that the old `objcopy
--only-keep-debug` command portage used to use is keeping enough
information around to gather local variables and backtraces. On the
crash server side, they consider that type of data as an upload type
of 'ELF' so let's just do that. If/when portage starts actually
splitting debug info from the binaries we'll possibly need to revisit
this CL so that we can continue to upload the full binaries to crash.
The main advantage of uploading the full binaries (i.e. ones with code)
is that we can disassemble the code in the stacktraces on the server.

BUG=b:171334299
TEST=go test
TEST=./upload_debug_symbols upload -dry-run=true -gs-path=gs://chromeos-image-archive/brya-release/R117-15544.0.0 -data-type=splitdebug

Change-Id: Ia296b0660e03e1b30360694717419f88a78cccb7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5260693
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#62939}