siso: fix phony dir handling
in chromium/mac,ios build, phony target is used for
top dir of frameworks etc and the target is used
as source of copy_bundle_data.

Siso ignored phony target as input
https://ninja-build.org/manual.html#_the_literal_phony_literal_rule
Then it would wrongly copy empty data to destination directory.

To match with ninja's behavior, use actual dir/file
if exists even if it is the phony target.

Bug: b/323778998, b/319585576
Change-Id: I001edb4a06a1f4fa517a6bcd04e6dae53ec27ea6