[golangbuild] log discovered modules, return early if no modules
If we discover there are no modules to test (e.g., a newly created x/
repo before it has Go packages added to it, or the luci-config branch
in x/build, and so on), stop early. Also log which modules were found.

This prevents the meaningless empty "fetch dependencies" step from
showing up in 'Steps & Logs'. It also makes it so that the code below
doesn't need to concern itself with the case of len(modules) == 0.
Most of it handles it just fine, but compileTestsInParallel panicked
due to a divide by zero.

I don't expect we would add something to fetchSubrepoAndRunTests that
provides a signal even in the case of no testable modules, but if we do,
we'll handle it accordingly then.

Change-Id: I8a6310609f19abad0b1c0ac36b3acf432b935707
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5281577
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#63053}