[golangbuild] handle step.Log.Write errors
We've started to check for errors from writing to step logs by default.
That should never fail, and this will notify us if that isn't the case.
Add it to the remaining few instances that haven't been updated yet to
avoid the old pattern from proliferating.

Change-Id: I9738ab921d3fb92c173e53e0bd049d8167b187c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5281581
Commit-Queue: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#63054}