[cloudbuildhelper] "Relocate" static paths in app.yaml.
In modules mode GAE considers the root of the **go module** to be the
application root. This is different from GOPATH mode, where the main
package directory (the one with app.yaml) is the application root.

The application root is significant in two ways:
  1. It is cwd of the server process. This is important if the server
     loads any files at runtime (e.g. HTML templates) from cwd.
  2. It is the root path used when evaluating static files patterns in
     app.yaml.

All our existing app YAMLs are written for GOPATH mode. And arguably
this is their truly correct form (they look just like one would
expect, e.g. like Python GAE YAMLs, with no spooky action at a
distance). So here we rewrite paths in them to be relatively to
the module root to make them compatible with modules mode.

Note the GAE team considers this discrepancy between modes
"Working as Intended": b/323980048.

This replaces the previous workaround that was trying to move
files instead (and dealing with resulting collisions). That was
a dumb thing to do.

R=bryner@google.com, fancl@google.com

Change-Id: I44c9a65cce158ee79c7b24aaad3703d5dd202ebe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5281555
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#63076}