upload_debug_symbols: Remove vmlinux step
The vmlinux step uploads the kernel vmlinux as an ELF type binary to the
crash servers. We already do that with the splitdebug step, so let's
just rely on the splitdebug logic for this. Remove the vmlinux related
code to deduplicate logic and make the code smaller.

BUG=b:322824656
TEST=go test

Change-Id: Ibfdb3a710e51135e51f180940a1d344a4f9cae43
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5297001
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63197}