[golangbuild] add a way to only set up a builder's environment
This change adds a new execution mode to golangbuild which installs
tools and exports environment variables in the same way that it would
for a particular builder. The execution mode skips the usual luciexe
protocol and does the setup directly.

The main use-case for this mode is making the gomote environment more
aligned with a builder's environment. Another use-case, however, is to
make it easier to create the builder's environment locally for
debugging.

Discussion:

The main reason we skip the luciexe protocol here is because it's more
complex than we need and less ergonomic. Most of the luciexe protocol
to us is about presentation and logging, but in the gomote scenario
that this CL is targeting, neither of those are very important.

We could maybe get away with using LUCIEXE_FAKEBUILD instead but that
is substantially less ergonomic, requiring the gomote server and/or a
user to construct a fake build proto. It's much simpler to specify
the builder name and just have the binary contact buildbucket itself.

Change-Id: I24c945eab91532881e573135dd65f4ee03841859
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5304135
Reviewed-by: Carlos Amedee <amedee@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#63239}