kajiya: fix "Access is denied" when storing blobs
This change fixes a bug where the blobstore would fail to store blobs
on Windows with an "Access is denied" error. The problem was that on
Windows, renaming a file to an existing file could fail not with an
os.ErrExist error, but with an os.ErrPermission error in case the blob
was already being used by an action, or we raced with another upload
on the rename.

This change fixes the issue by additionally checking for the existence
of the destination file if the rename fails and treating it as a
successful rename in that case.

Bug: b/323454226
Change-Id: I1863cc20b0242580d0a6977aeac296a941010dbe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5312842
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Philipp Wollermann <philwo@chromium.org>
Auto-Submit: Philipp Wollermann <philwo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63264}