[golangbuild] add an explicit test timeout scale property
This change adds an explicit test timeout scale property that sets both
the GO_TEST_TIMEOUT_SCALE environment variable (for go tool dist) and
the -timeout flag to go test.

This change aligns golangbuild more closely with the legacy
infrastructure. Without it, we do not apply any timeout scaling for
subrepos, which is important specifically for longtest subrepo builders.

Change-Id: I9fc61396134057228848701b8f3dd7aae8633ac5