[golangbuild] set and use GOPLSCACHE
GOPLSCACHE exists to point gopls (and its tests) at a specific directory
for its cache. This is necessary on builders where the cache might be
set to a more permanent location, causing it to grow over time and eat
up the machine's disk space.

This change creates an explicit directory for GOPLSCACHE next to GOCACHE
in the working directory. This will ensure the directory gets deleted
every run. We do this instead of TMPDIR for consistency, and so the
cache directory is easily recognizable.

For go.dev/issue/65725.

Change-Id: I1b6a58de5b40133640e5860eb3d2549d1a2067d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5314212
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Cr-Commit-Position: refs/heads/main@{#63283}