Add older WebOb wheel
Adds a wheel for WebOb 1.5.0, which is the stable version of the
pre-release version currently used by the Catapult project. Providing this wheel will allow Catapult to switch to the vpython-provided
version instead of having a local copy, which will in turn allow us
to more easily upgrade the version when we switch to Python 3.11.

Bug: chromium:324114898
Change-Id: Ia005b903cc089d15e06faf484db057e55e9e5360
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5322084
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63350}