Remove tzeller@google.com from go/src/infra/fleet/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/tzeller@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iffac68b0c85de047ac9d7a04e14006b96611e8c3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5347190
Reviewed-by: Luke Tidd <luket@google.com>
Commit-Queue: Luke Tidd <luket@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#63596}