upload_debug_symbols: Don't try to upload missing splitdebug files
Sometimes the splitdebug file can't be found, because it doesn't have a
debug identifier like the build ID in it, or because the name of the
symbol doesn't exactly match what breakpad used to generate the symbols.
In this case, we won't upload the splitdebug file to crash, but we'll
still try to upload the binary if the breakpad file exists.

Fix this by only adding a splitdebug upload task if the file actually
exists.

BUG=b:328215310
TEST=go test

Change-Id: I395e0a4fe3c1a305071e2e7b3044de76012f0567
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5354303
Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#63665}