device-manager: implement LeaseDevice rpc
LeaseDevice executes a lease request in a transaction. In this
iteration, the request must contain a specific request for a device. The
RPC will check for availability, lease the device, and update its state.

The RPC also checks for request idempotency. It will return an unexpired
lease record if found.

BUG=b/326503964
TEST=make test; manual grpcurl lease requests

Change-Id: I54b12d46ca7f55693f5d0052590b3e9f1ead5e75
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5324212
Reviewed-by: Ashish Gandhi <ashishgandhi@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#63740}