device-manager: add ExtendLeaseRequests table
This table will manage the ExtendLease requests. A record will be
created every time the RPC is called. The table is append-only from the
service. The records will eventually be sent to BQ and they will only
persist in this table for 3 days after the request_time.

BUG=b/326503718
TEST=alembic upgrade head

Change-Id: I0ea7313ffc414cf81739edc8d198d4ea9f5169c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5350450
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Ashish Gandhi <ashishgandhi@google.com>
Cr-Commit-Position: refs/heads/main@{#63745}