[isolation] [owners-cleanup] Add ayatane and renewitt as OWNERS for isolation jail
This CL which landed in the recent past cleans up some of the OWNERS.

https://chromium-review.googlesource.com/c/infra/infra/+/5368674

As part of that CL I became an OWNER of nsjail because I landed a CL in the
project a while back. The CL that I landed does nothing. I grepped the code for
string-valued variables that could be made into local constants and stumbled
across this project.

https://chromium-review.googlesource.com/c/infra/infra/+/3880895

BUG=None
TEST=None

Change-Id: I3dcca1dc39c048a898793ce5a1a622e9de251fca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/5368253
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#63757}