[golangbuild] add shard-by-weight experiment for test sharding
This change adds the golang.shard_by_weight experiment which uses
predetermined weights on `go tool dist test` tests to shard them more
evenly (in terms of the amount of work that needs to happen).

It adds the testweights package to contain this information as well as a
generator script to regenerate the weights from the local execution of
the tests. The weights are determined as the number of seconds it took
the test to execute, since it seems all fast tests always take at least
around a second to build.

This isn't perfect, but it's a start. This change also gives us a way to
keep tweaking this until we get it right.

Change-Id: I8163f01c69badef69e804c0c2ce7a46c315b7544